perf(studio): isolate default layout implementation
