+++ packages/sanity/src/core/studio/__tests__/StudioLayout.test.tsx
@@ -4,7 +4,7 @@ import {type ReactNode} from 'react'
 import {afterEach, beforeEach, describe, expect, it, vi} from 'vitest'
 
 import {type StudioReadyMeasured as StudioReadyMeasuredType} from '../__telemetry__/bootstrap.telemetry'
-import {type StudioLayoutComponent as StudioLayoutComponentType} from '../StudioLayoutComponent'
+import {type StudioLayoutComponent as StudioLayoutComponentType} from '../StudioLayoutComponentImpl'
 
 // StudioLayout renders `@sanity/ui` components that require a `ThemeProvider`
 // via context. Wrap every render in a minimal studio theme.
@@ -111,7 +111,7 @@ describe('StudioLayoutComponent telemetry', () => {
 
     const {useTelemetry} = await import('@sanity/telemetry/react')
     ;(useTelemetry as ReturnType<typeof vi.fn>).mockReturnValue({log: telemetryLog})
-    ;({StudioLayoutComponent} = await import('../StudioLayoutComponent'))
+    ;({StudioLayoutComponent} = await import('../StudioLayoutComponentImpl'))
     ;({StudioReadyMeasured} = await import('../__telemetry__/bootstrap.telemetry'))
   })
